feat: utilize blueprints for each grouped logic
